+++ src/onegov.election_day/onegov/election_day/layouts/__init__.py
@@ -5,18 +5,19 @@ from onegov.election_day.layouts.election_compound import \
 from onegov.election_day.layouts.mail import MailLayout
 from onegov.election_day.layouts.manage import ManageDataSourceItemsLayout
 from onegov.election_day.layouts.manage import ManageDataSourcesLayout
-from onegov.election_day.layouts.manage import ManageElectionsLayout
 from onegov.election_day.layouts.manage import ManageElectionCompoundsLayout
+from onegov.election_day.layouts.manage import ManageElectionsLayout
 from onegov.election_day.layouts.manage import ManageLayout
 from onegov.election_day.layouts.manage import ManageSubscribersLayout
+from onegov.election_day.layouts.manage import ManageUploadTokensLayout
 from onegov.election_day.layouts.manage import ManageVotesLayout
 from onegov.election_day.layouts.vote import VoteLayout
 
 
 __all__ = [
     'DefaultLayout',
-    'ElectionLayout',
     'ElectionCompoundLayout',
+    'ElectionLayout',
     'MailLayout',
     'ManageDataSourceItemsLayout',
     'ManageDataSourcesLayout',
@@ -24,6 +25,7 @@ __all__ = [
     'ManageElectionsLayout',
     'ManageLayout',
     'ManageSubscribersLayout',
+    'ManageUploadTokensLayout',
     'ManageVotesLayout',
     'VoteLayout',
 ]