+++ src/onegov.election_day/onegov/election_day/tests/collections/test_upload_token_collection.py
@@ -1,28 +1,18 @@
 from onegov.election_day.collections import UploadTokenCollection
-from uuid import uuid4
 
 
 def test_upload_token_collection(session):
 
     collection = UploadTokenCollection(session)
-    assert collection.list() == []
+    assert collection.query().all() == []
 
     token = collection.create()
-    assert collection.list() == [token]
+    assert collection.query().all() == [token]
 
-    token = collection.create(token=token)
-    assert collection.list() == [token]
+    assert collection.by_id(token.id) == token
 
-    another_token = collection.create(token=uuid4())
-    assert set(collection.list()) == set([token, another_token])
+    another_token = collection.create()
+    assert set(collection.query().all()) == set([token, another_token])
 
     collection.delete(token)
-    assert collection.list() == [another_token]
-
-    collection.create()
-    collection.create()
-    collection.create()
-    assert len(collection.list()) == 4
-
-    collection.clear()
-    assert collection.list() == []
+    assert collection.query().all() == [another_token]